feat: params modified with hook
