+++ src/lambdas/axios-request-success.lambda.ts
@@ -6,5 +6,6 @@ export const axiosRequestSuccessLambda = (
   // Do something before request is sent
   config.url = `http://localhost:8000/api` + config.url
   console.log(`Requesting to URL: ` + config.url)
+  console.log(`Params: ` + JSON.stringify(config.params))
   return config
 }
